+++ resources/sql/hotel-room.sql
@@ -1,7 +1,7 @@
 drop table if exists chr_hotel;
 create table chr_hotel
 (
-    hotel_id      int primary key AUTO_INCREMENT,
+    hotel_id      bigint(20) primary key AUTO_INCREMENT,
     province      varchar(100) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci not null,
     city          varchar(100) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci not null,
     district      varchar(100) CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci not null,
@@ -16,7 +16,7 @@ create table chr_hotel
     parking       boolean                                                       not null,
     spa           boolean                                                       not null,
     chess_room    boolean                                                       not null,
-    star_level    int                                                           not null
+    star_level    bigint(20)                                                           not null
 ) ENGINE = InnoDB
   AUTO_INCREMENT = 1
   CHARACTER SET = utf8mb4
@@ -26,8 +26,8 @@ create table chr_hotel
 drop table if exists chr_room_type;
 create table chr_room_type
 (
-    type_id    int primary key AUTO_INCREMENT,
-    hotel_id   int            not null,
+    type_id    bigint(20) primary key AUTO_INCREMENT,
+    hotel_id   bigint(20)            not null,
     price      decimal(18, 2) not null,
     breakfast  boolean        not null,
     windows    boolean        not null,
@@ -43,11 +43,11 @@ create table chr_room_type
 drop table if exists chr_room;
 create table chr_room
 (
-    room_id     int primary key,
-    hotel_id    int not null,
-    room_number int not null,
-    type_id     int not null,
-    layout_id   int not null
+    room_id     bigint(20) primary key,
+    hotel_id    bigint(20) not null,
+    room_number bigint(20) not null,
+    type_id     bigint(20) not null,
+    layout_id   bigint(20) not null
 ) ENGINE = InnoDB
   AUTO_INCREMENT = 1
   CHARACTER SET = utf8mb4
@@ -57,14 +57,14 @@ create table chr_room
 drop table if exists chr_hotel_picture;
 create table chr_hotel_picture
 (
-    hotel_id     int not null,
+    hotel_id     bigint(20) not null,
     picture_path varchar(200)
 );
 
 drop table if exists chr_room_type_picture;
 create table chr_room_type_picture
 (
-    type_id      int not null,
+    type_id      bigint(20) not null,
     picture_path varchar(200)
 ) ENGINE = InnoDB
   AUTO_INCREMENT = 1
@@ -75,9 +75,9 @@ create table chr_room_type_picture
 drop table if exists chr_layout;
 create table chr_layout
 (
-    layout_id        int primary key AUTO_INCREMENT,
-    hotel_id         int not null,
-    floor            int not null,
+    layout_id        bigint(20) primary key AUTO_INCREMENT,
+    hotel_id         bigint(20) not null,
+    floor            bigint(20) not null,
     plane_graph_path varchar(200)
 ) ENGINE = InnoDB
   AUTO_INCREMENT = 1
